pyproject.toml
[project]
name = "superset-engine-d1"
version = "0.1.0"
description = ""
authors = [
    {name = "Chad Rossouw",email = "chadrossouw7247@gmail.com"},
    {name = "Shreyas Rao",email = "raoshreyas2004@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"

[tool.poetry]
packages = [{include = "superset_engine_d1", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
apache-superset=">=3.0"
sqlalchemy-d1 = { path = "../sqlalchemy-d1" }
marshmallow = "3.26.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# Allows auto-discoverability by superset
[project.entry-points."superset.db_engine_specs"]
d1 = "superset_engine_d1.d1_engine_spec:D1EngineSpec"


[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "mypy (>=1.18.2,<2.0.0)"
]
